First step in refactoring the Irradation Report
