+++ include/Analyzer.h
@@ -49,6 +49,7 @@
 #include <TAxis.h>
 #include <TCanvas.h>
 
+#include <AnalyzerTools.hh>
 
 namespace insur {
 
@@ -86,7 +87,7 @@ namespace insur {
   typedef std::map<int, TrackCollection> TrackCollectionMap;
 
 
-  class Analyzer {
+  class Analyzer : private AnalyzerTools {
   public:
     Analyzer();
     virtual ~Analyzer() {}
@@ -178,7 +179,6 @@ namespace insur {
     void analyzeGeometry(Tracker& tracker, int nTracks = 1000);
     void computeBandwidth(Tracker& tracker);
     void computeTriggerFrequency(Tracker& tracker);
-    void computeIrradiationPowerConsumption(Tracker& tracker);
     void analyzePower(Tracker& tracker);
     void createGeometryLite(Tracker& tracker);
     TH2D& getMapPhiEta() { return mapPhiEta; }
@@ -247,8 +247,6 @@ namespace insur {
     std::map<std::string, SummaryTable>& getTriggerEfficiencySummaries() { return triggerEfficiencySummaries_; }
     std::map<std::string, SummaryTable>& getTriggerPuritySummaries() { return triggerPuritySummaries_; }
     std::map<std::string, SummaryTable>& getTriggerDataBandwidthSummaries() { return triggerDataBandwidthSummaries_; }
-    std::map<std::string, SummaryTable>& getSensorsIrradiationPowerSummary() { return sensorsIrradiationPowerSummary_; }
-    std::map<std::string, SummaryTable>& getSensorsIrradiationSummary() { return sensorsIrradiationSummary_; }
     
     double getTriggerPetalCrossoverR() const { return triggerPetalCrossoverR_; }
     const std::pair<Circle, Circle>& getSampleTriggerPetal() const { return sampleTriggerPetal_; }
@@ -338,8 +336,6 @@ namespace insur {
     std::map<std::string, SummaryTable> triggerFrequencyTrueSummaries_, triggerFrequencyFakeSummaries_, triggerFrequencyMisfilteredSummaries_, triggerFrequencyCombinatorialSummaries_, triggerFrequencyInterestingSummaries_;
     std::map<std::string, SummaryTable> triggerRateSummaries_, triggerEfficiencySummaries_, triggerPuritySummaries_;
     std::map<std::string, SummaryTable> triggerDataBandwidthSummaries_;
-    std::map<std::string, SummaryTable> sensorsIrradiationPowerSummary_;
-    std::map<std::string, SummaryTable> sensorsIrradiationSummary_;
 
     std::map<std::string, SummaryTable> stripOccupancySummaries_;
     std::map<std::string, SummaryTable> hitOccupancySummaries_;
@@ -428,7 +424,6 @@ namespace insur {
     //void fillPowerMap(Tracker& tracker);
     void clearMaterialBudgetHistograms();
     void prepareTriggerPerformanceHistograms(const int& nTracks, const double& etaMax, const vector<double>& triggerMomenta, const vector<double>& thresholdProbabilities);
-    void preparePowerHistograms();
     void prepareTriggerProcessorHistograms();
     void clearGeometryHistograms();
     void clearCells();
@@ -455,8 +450,6 @@ namespace insur {
     Color_t lastPickedColor;
     int geometryTracksUsed;
     int materialTracksUsed;
-    void prepareTrackerMap(TH2D& myMap, const std::string& name, const std::string& title);
-    void prepareRadialTrackerMap(TH2D& myMap, const std::string& name, const std::string& title);
     void fillAvailableSpacing(Tracker& tracker, std::vector<double>& spacingOptions);
     static constexpr double maximum_n_planes = 13.;
